Added find/erase functionality to CellList via iterators. Moved extra functions from ComponentMap into IncrementalUpdate.hpp.
