+++ src/CellList.hpp
@@ -39,20 +39,35 @@ template <typename GraphType>
 struct CellList
 {
 
+    //TODO: maybe we need a multidimensional iterator, to iterate over the outer and inner
     //CellListData data;
     using ComponentMapType = ComponentMap<typename GraphType::key_type>;
     using ComponentMapIter = typename ComponentMapType::iterator;
     using CellListObjectType = typename ComponentMapType::key_type;
     using CellListDataType = std::vector<std::vector<CellListObjectType>>;
     //using GraphType = YAGL::Graph<Plant::mt_key_type, Plant::MT_NodeData>;
+
+    using outer_iterator = typename CellListDataType::iterator;
+    using inner_iterator = typename std::vector<CellListObjectType>::iterator;
+
+    struct IteratorPair
+    {
+        typename CellListDataType::iterator outer_iterator;
+        typename std::vector<CellListObjectType>::iterator inner_iterator;
+    };
+
+    using iterator_pair = IteratorPair;
+
     CellListDataType data;
 
     CartesianGrid2D grid;
-    GraphType& graph;
-    ComponentMapType& component_matches;
+    GraphType* graph;
+    ComponentMapType* component_matches;
     int cell_range;
 
-    CellList(CartesianGrid2D& _grid, GraphType& _graph, ComponentMapType& _component_matches)
+    CellList() = default;
+
+    CellList(CartesianGrid2D& _grid, GraphType* _graph, ComponentMapType* _component_matches)
         : grid(_grid), graph(_graph), component_matches(_component_matches), data(_grid.totalNumCells())
     {
         //TODO make the cell size ratio a paramater 
@@ -61,7 +76,7 @@ struct CellList
 
         //build the cell list from the system graph 
         std::cout << "Building cell list\n";
-        for(auto& item : component_matches) insert_match(item);
+        for(auto it = component_matches->begin(); it != component_matches->end(); it++) insert_match(it);
         std::cout << "Cell list has been built\n";
         std::cout << "The cell list has " << totalNumCells() << " reaction cells\n"; 
         std::cout << "Components mapped: " << getTotalSize() << "\n";
@@ -78,26 +93,79 @@ struct CellList
         for(const auto& cell : data) sum += cell.size();
         return sum;
     }
-    template<typename T> //Temporary fix since ComponentMapIter throws an error
-    std::size_t locate_cell(T& comp_iter)
+
+    std::size_t locate_cell(ComponentMapIter& comp_iter)
     {
-        auto& match = comp_iter.second;
-        auto& anchor_data = graph.findNode(match.anchor)->second.getData();
+        auto& match = comp_iter->second;
+        auto& anchor_data = graph->findNode(match.anchor)->second.getData();
         double xp = anchor_data.position[0];
         double yp = anchor_data.position[1];
-        auto key = comp_iter.first;
+        auto key = comp_iter->first;
         int ic, jc;
         grid.locatePoint(xp, yp, ic, jc);
         return grid.cardinalCellIndex(ic, jc);
     }
 
-    template<typename T> //Temporary fix since ComponentMapIter throws an error
-    void insert_match(T& comp_iter)
+    void insert_match(ComponentMapIter& comp_iter)
     {
         auto cardinal = locate_cell(comp_iter);
-        data[cardinal].push_back(comp_iter.first);
+        data[cardinal].push_back(comp_iter->first);
     }
-    
+
+
+    //The find function must search the current cell and potentially it's neigbhors to see where a component is
+    iterator_pair find(ComponentMapIter& comp_iter)
+    {
+        //first see if the component is in the cell it would be currently mapped to
+        auto c = locate_cell(comp_iter);
+        bool found = false;
+        for(auto i = 0; i < data[c].size(); i++)
+        {
+            if(comp_iter->first == data[c][i]) {
+                found = true;
+                std::cout << "Found component " << comp_iter->first << " in cell " << c << " at index " << i << "\n";
+                auto outIter = data.begin()+c;
+                auto inIter = outIter->begin()+i;
+                return iterator_pair{outIter, inIter};
+            }
+        }
+
+        if(!found) //expand the search, since currently mapped cell, may not be the original (components move)
+        {
+            int imin, imax, jmin, jmax;
+            getCells(c, imin, imax, jmin, jmax);
+            for(auto i = imin; i < imax; i++)
+            {
+                for(auto j = jmin; j < jmax; j++)
+                {
+                    auto d = grid.cardinalCellIndex(i, j);
+                    //if(d == c) continue;
+                    for(auto k = 0; k < data[d].size(); k++)
+                    {
+                        if(comp_iter->first == data[c][k]) {
+                            found = true;
+                            std::cout << "Found component " << comp_iter->first << " in cell " << c << " at index " << k << "\n";
+                            auto outIter = data.begin()+c;
+                            auto inIter = outIter->begin()+i;
+                            return iterator_pair{outIter, inIter};
+                        }
+                    }
+                }
+            }
+        }
+        //else we return that we didn't find anything
+        auto outIter = data.end();
+        auto inIter = outIter->end();
+        return iterator_pair {outIter, inIter};
+    }
+
+    void erase(ComponentMapIter& comp_iter)
+    {
+        auto r = find(comp_iter);
+        if(r.outer_iterator != data.end())
+            r.outer_iterator->erase(r.inner_iterator);
+    }
+
     std::size_t totalNumCells() { return grid.totalNumCells(); }
     
     std::size_t cardinalCellIndex(int i, int j) { return grid.cardinalCellIndex(i, j); }
